+++ docs/examples/defaultlabelprovider/translations.tsx
@@ -111,7 +111,7 @@ const labels = {
     tooltipMessage: myI18nTranslator('Click to learn more'),
   },
   Toast: {
-    accessibilityDismissButtonLabel: myI18nTranslator('Dismiss toast'),
+    accessibilityDismissButtonLabel: myI18nTranslator('Dismiss message'),
     accessibilityIconSuccessLabel: myI18nTranslator('Success message'),
     accessibilityIconErrorLabel: myI18nTranslator('Error message'),
     accessibilityProcessingLabel: myI18nTranslator('Processing message'),