+++ packages/gestalt/src/Checkbox/InternalCheckbox.tsx
@@ -9,6 +9,7 @@ import FormErrorMessage from '../sharedSubcomponents/FormErrorMessage';
 import FormHelperText from '../sharedSubcomponents/FormHelperText';
 import Text from '../Text';
 import useFocusVisible from '../useFocusVisible';
+import useInExperiment from '../useInExperiment';
 import useInteractiveStates from '../utils/useInteractiveStates';
 
 type Props = {
@@ -85,8 +86,11 @@ const InternalCheckboxWithForwardRef = forwardRef<HTMLInputElement, Props>(funct
     handleOnMouseLeave,
     handleOnBlur,
     handleOnFocus,
+    handleOnMouseDown,
+    handleOnMouseUp,
     isFocused,
     isHovered,
+    isActive,
   } = useInteractiveStates();
   const { isFocusVisible } = useFocusVisible();
 
@@ -100,31 +104,76 @@ const InternalCheckboxWithForwardRef = forwardRef<HTMLInputElement, Props>(funct
     ariaDescribedby = `${id}-helperText`;
   }
 
-  let bgStyle = styles.enabled;
-  if (disabled) {
-    bgStyle = styles.disabled;
-  } else if (checked || indeterminate) {
-    bgStyle = styles.checked;
-  }
+  const isInVRExperiment =
+    useInExperiment({
+      webExperimentName: 'web_gestalt_visualRefresh',
+      mwebExperimentName: 'web_gestalt_visualRefresh',
+    }) && !style;
 
-  let borderStyle = styles.border;
-  if (disabled) {
-    borderStyle = styles.borderDisabled;
-  } else if (!disabled && (checked || indeterminate)) {
-    borderStyle = styles.borderSelected;
-  } else if (errorMessage) {
-    borderStyle = styles.borderError;
-  } else if (!disabled && isHovered) {
-    borderStyle = styles.borderHovered;
-  }
+  const iconSizes = {
+    sm: 8,
+    md: 12,
+  };
+
+  const vrIconSizes = {
+    sm: 12,
+    md: 16,
+  };
 
   const borderRadiusStyle = size === 'sm' ? styles.borderRadiusSm : styles.borderRadiusMd;
-
+  const borderRadiusStyleVr = size === 'sm' ? styles.vrBorderRadiusSm : styles.vrBorderRadiusMd;
   const styleSize = size === 'sm' ? styles.sizeSm : styles.sizeMd;
+  const vrTextColor = disabled ? 'disabled' : undefined;
+  const textColor = disabled ? 'subtle' : undefined;
+  const vrIconColorEnabled = errorMessage ? 'error' : 'inverse';
+  const vrIconColor = disabled ? 'disabled' : vrIconColorEnabled;
+  const unchecked = !(checked || indeterminate);
 
-  const divStyles = classnames(bgStyle, borderStyle, borderRadiusStyle, styleSize, styles.check, {
-    [focusStyles.accessibilityOutlineFocus]: isFocused && isFocusVisible,
+  const bgStyle = classnames({
+    [styles.enabled]: !isInVRExperiment && !disabled && unchecked,
+    [styles.disabled]: disabled,
+    [styles.checked]: !unchecked && !disabled,
+    [styles.error]: isInVRExperiment && errorMessage && unchecked,
+    [styles.errorChecked]: isInVRExperiment && errorMessage && !unchecked,
+    [styles.hovered]:
+      isInVRExperiment && !disabled && isHovered && !isActive && !errorMessage && !unchecked,
+    [styles.hoveredError]:
+      !disabled && isHovered && !isActive && isInVRExperiment && errorMessage && !unchecked,
+    [styles.pressed]: isInVRExperiment && isActive && !disabled && !unchecked && !errorMessage,
+    [styles.pressedError]: isInVRExperiment && isActive && !disabled && !unchecked && errorMessage,
   });
+
+  const borderStyle = classnames({
+    [styles.border]: !disabled && unchecked && !errorMessage && !isHovered,
+    [styles.borderDisabled]: disabled,
+    [styles.borderSelected]: !disabled && !unchecked,
+    [styles.borderErrorUnchecked]: errorMessage && unchecked,
+    [styles.borderErrorChecked]: errorMessage && !unchecked && isInVRExperiment,
+    [styles.borderHovered]:
+      !disabled && isHovered && !isActive && unchecked && (isInVRExperiment || !errorMessage),
+    [styles.borderHoveredError]:
+      !disabled && isHovered && !isActive && isInVRExperiment && errorMessage && unchecked,
+    [styles.borderPressed]: isInVRExperiment && isActive && !disabled && unchecked && !errorMessage,
+    [styles.borderPressedError]:
+      isInVRExperiment && isActive && !disabled && unchecked && errorMessage,
+  });
+
+  const divStyles = classnames(
+    bgStyle,
+    borderStyle,
+    isInVRExperiment ? borderRadiusStyleVr : borderRadiusStyle,
+    styleSize,
+    styles.check,
+    {
+      [styles.thickBorder]:
+        !isInVRExperiment || (errorMessage && unchecked) || isHovered || isActive,
+      [styles.thinBorder]:
+        isInVRExperiment &&
+        !((isFocused && isFocusVisible) || isHovered || isActive || errorMessage),
+      [focusStyles.accessibilityOutlineFocus]: isFocused && isFocusVisible && !isInVRExperiment,
+      [styles.focus]: isFocused && isFocusVisible && isInVRExperiment,
+    },
+  );
 
   const inputStyles = classnames(styles.input, styleSize, {
     [styles.inputEnabled]: !disabled,
@@ -151,17 +200,19 @@ const InternalCheckboxWithForwardRef = forwardRef<HTMLInputElement, Props>(funct
             // @ts-expect-error - TS2322 - Type '(event: React.ChangeEvent<HTMLInputElement>) => void' is not assignable to type 'MouseEventHandler<HTMLInputElement>'.
             onClick={handleClick}
             onFocus={handleOnFocus}
+            onMouseDown={handleOnMouseDown}
             onMouseEnter={handleOnMouseEnter}
             onMouseLeave={handleOnMouseLeave}
+            onMouseUp={handleOnMouseUp}
             type="checkbox"
           />
           <div className={divStyles} style={style}>
             {(checked || indeterminate) && (
               <Icon
                 accessibilityLabel=""
-                color="inverse"
+                color={isInVRExperiment ? vrIconColor : 'inverse'}
                 icon={indeterminate ? 'dash' : 'check'}
-                size={size === 'sm' ? 8 : 12}
+                size={isInVRExperiment ? vrIconSizes[size] : iconSizes[size]}
               />
             )}
           </div>
@@ -177,17 +228,26 @@ const InternalCheckboxWithForwardRef = forwardRef<HTMLInputElement, Props>(funct
           >
             <Label htmlFor={id}>
               <Box paddingX={1}>
-                <Text color={disabled ? 'subtle' : undefined} size={size === 'sm' ? '200' : '300'}>
+                <Text
+                  color={isInVRExperiment ? vrTextColor : textColor}
+                  size={size === 'sm' ? '200' : '300'}
+                >
                   {label}
                 </Text>
               </Box>
             </Label>
             <Box paddingX={1}>
-              {helperText && !errorMessage ? (
-                <FormHelperText id={`${id}-helperText`} size={size} text={helperText} />
+              {helperText ? (
+                <FormHelperText
+                  disabled={disabled}
+                  id={`${id}-helperText`}
+                  noPadding
+                  size={size}
+                  text={helperText}
+                />
               ) : null}
               {errorMessage ? (
-                <FormErrorMessage id={`${id}-error`} size={size} text={errorMessage} />
+                <FormErrorMessage id={`${id}-error`} noPadding size={size} text={errorMessage} />
               ) : null}
             </Box>
           </Box>